More grid snapping options, ability to position tracks beyond screen bounds

+++ Assets/Scripts/UI/SnapGrid.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class SnapGrid : UIElement {
     public float lastSnapGridSetting;
-    public int lastBPM;
+    public float lastBPM;
     public bool despawned;
 
     public SnapGrid() {
